# public/openapi.yaml
openapi: 3.0.3
info:
  title: pNode Pulse API
  description: |
    Public REST API for accessing pNode network data.

    ## Authentication

    Most endpoints are public and rate-limited. For higher limits, use an API key:

    ```
    Authorization: Bearer pk_your_api_key
    ```

    ## Rate Limits

    | Tier | Requests/min | Requests/day |
    |------|--------------|--------------|
    | Anonymous | 30 | 1,000 |
    | Free API Key | 60 | 10,000 |
    | Premium | 300 | 100,000 |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Max requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  version: 1.0.0
  contact:
    name: pNode Pulse Support
    url: https://pulse.rectorspace.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://pulse.rectorspace.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Development
...
paths:
...
  /nodes:
    get:
      tags:
        - Nodes
      summary: List all nodes
      description: Returns a paginated list of all pNodes with optional filtering.
      operationId: listNodes
      parameters:
        - name: status
          in: query
          description: Filter by node status
          schema:
            type: string
            enum: [all, active, inactive]
            default: all
        - name: version
          in: query
          description: Filter by software version
          schema:
            type: string
        - name: search
          in: query
          description: Search by IP address or public key
          schema:
            type: string
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: orderBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [lastSeen, firstSeen, address, version, isActive]
            default: lastSeen
        - name: order
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated list of nodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeList'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /leaderboard:
    get:
      tags:
        - Leaderboard
      summary: Get node leaderboard
      description: Returns top or bottom performing nodes by specified metric.
      operationId: getLeaderboard
      parameters:
        - name: metric
          in: query
          description: Metric to rank by
          schema:
            type: string
            enum: [uptime, cpu, ram, storage]
            default: uptime
        - name: order
          in: query
          description: Ranking order
          schema:
            type: string
            enum: [top, bottom]
            default: top
        - name: limit
          in: query
          description: Number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: period
          in: query
          description: Time period for metrics
          schema:
            type: string
            enum: [24h, 7d, 30d, all]
            default: 7d
      responses:
        '200':
          description: Leaderboard rankings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
...
components:
  schemas:
...
    NodeList:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    Node:
      type: object
      properties:
        id:
          type: integer
        address:
          type: string
          example: "192.190.136.36:9001"
        pubkey:
          type: string
          nullable: true
        version:
          type: string
          example: "0.6.0"
        isActive:
          type: boolean
        lastSeen:
          type: string
          format: date-time
        firstSeen:
          type: string
          format: date-time
...
    Leaderboard:
      type: object
      properties:
        metric:
          type: string
        period:
          type: string
        order:
          type: string
        rankings:
          type: array
          items:
            type: object
            properties:
              rank:
                type: integer
              nodeId:
                type: integer
              address:
                type: string
              version:
                type: string
              value:
                type: number
              change:
                type: integer
                description: Rank change from previous period
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests. Please try again later.
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: Optional API key for higher rate limits
security:
  - {}
  - ApiKeyAuth: []